chore: refactor autorun/reaction wrappers
